requirements.txt
psycopg2-binary
requests
logging
